--- src/pages/background/constants/glossary.js
@@ -1,41 +1,0 @@
-import React from 'react';
-
-const glossaryInfo = [
-  {
-    description: (
-      <ul className="glossary-list">
-        <li>
-          {/* <img src={} alt={} /> */}
-          <p><strong>AEWA,</strong><a href="https://www.unep-aewa.org/en"> African-Eurasian Waterbird Agreement.</a></p>
-        </li>
-        <li>
-          {/* <img src={} alt={} /> */}
-          <p><strong>CAF,</strong><a href="https://www.cms.int/en/document/central-asian-flyway-action-plan-conservation-migratory-waterbirds-and-their-habitats"> Central Asian Flyway.</a></p>
-        </li>
-        <li>
-          {/* <img src={} alt={} /> */}
-          <p><strong>CMS,</strong><a href="https://www.cms.int/newsroom/?title=cop13"> Convention on Migratory Species.</a></p>
-        </li>
-        <li>
-          {/* <img src={} alt={} /> */}
-          <p><strong>CSR,</strong><a href="https://www.unep-aewa.org/en/document/7th-edition-conservation-status-report-csr7-including-annexes-1"> Conservation Status Review, a report made for AEWA.</a></p>
-        </li>
-        <li>
-          {/* <img src={} alt={} /> */}
-          <p><strong>EAAF Partnership,</strong><a href="https://www.eaaflyway.net/"> East Asian-Australasian Flyway Partnership.</a></p>
-        </li>
-        <li>
-          {/* <img src={} alt={} /> */}
-          <p><strong>EUBD,</strong><a href="https://ec.europa.eu/environment/nature/legislation/birdsdirective/index_en.htm"> European Union Birds Directive.</a></p>
-        </li>
-
-        <li>
-          {/* <img src={} alt={} /> */}
-          <p><strong>WHSRN,</strong><a href="https://www.whsrn.org/"> Western Hemisphere Shorebird Reserve Network.</a></p>
-        </li>
-      </ul>
-    )
-  }
-];
-
-export default glossaryInfo;
