+++ src/pages/background/waw/constants.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import Image from 'images/background-tab/duck.png';
 
+import './styles.scss';
+
 const WAWInfo = [
   {
     intro: (
