+++ src/pages/static-page/styles.scss
@@ -25,4 +25,9 @@
   p {
     line-height: 26px;
   }
+
+  .flex-container {
+    display: flex;
+    padding: 50px 0;
+  }
 }
